move environment util to logging package
